+++ scripts/omtk/core/classModule.py
@@ -378,3 +378,4 @@ class Module(object):
         else:
             return []
 
+
